Improve package_info fallbacks
Most importantly, this commit adds a fallback for when no
package_info.json file exists, as is the case for the pipsi venv when
installed through get-pipsi.py.  I also moved all the fallback code to
get_package_info.  In doing so, I improved the list_everything fallbacks
which prior to this change gave missing data (like None for version).  I
also believe that putting all the fallback code in one place will make
it easier to maintain.

Additionally, I changed list_everything to actually make use of the name
stored in package_info, which was currently unused.  There are probably
other places that could make use of it too, since if all the code just
goes by the venv name, the name data might as well not even exist.